{{ cookiecutter.project_name }}/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100

[tool.django-stubs]
django_settings_module = "{{ cookiecutter.package_name }}.settings"

[tool.flakeheaven]
exclude = [
  "*.pot",
  "*.py[co]",
  "*/env/*",
  "*/migrations/*",
  ".env",
  ".git",
  ".gitignore",
  "__pycache__",
  "venv",
]
format = "grouped"
max_line_length = 100
show_source = true

[tool.flakeheaven.exceptions."**/__init__.py"]
pyflakes = ["-F401", "-F403"]

[tool.flakeheaven.plugins]
mccabe = ["+*"]
pycodestyle = [
  "+*",
  "-E501",  # Ignore lines that are too long (Black will handle them, but it will ignore long strings, which is desired)
  "-W503", "-W504"  # Ignore line breaks around binary operators (Black will handle them)
]
pyflakes = ["+*"]
pylint = ["+*"]

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
color_output = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
error_summary = true
follow_imports = "skip"
follow_imports_for_stubs = true
ignore_missing_imports = true
no_implicit_optional = true
plugins = [
  "mypy_django_plugin.main"
]
pretty = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
warn_unused_ignores = true

[tool.poetry]
authors = ["{{ cookiecutter.author_full_name <{{ cookiecutter.author_email }}>"]
description = "{{ cookiecutter.project_description }}"
license = "MIT"
name = "{{ cookiecutter.package_name }}"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
celery = {extras = ["redis"], version = "^5.2.7"}
django = "^4"
django-allauth = "^0.51.0"
django-allauth-2fa = "^0.10.0"
django-celery-beat = "^2.4.0"
django-celery-results = "^2.4.0"
django-constance = {extras = ["database"], version = "^2.9.1"}
django-cors-headers = "^3.13.0"
django-debug-toolbar = "^3.8.1"
django-environ = "^0.9.0"
django-extensions = "*"
django-health-check = "^3.17.0"
django-htmx = "^1.13.0"
django-otp = "^1.1.4"
django-recaptcha = "^3.0.0"
django-redis = "^5.2.0"
django-storages = {extras = ["boto3"], version = "^1.13.1"}
django-timezone-field = "^5.0"
django-waffle = "^3.0.0"
gunicorn = {extras = ["setproctitle"], version = "^20.1.0"}
ipython = "*"
psutil = "^5.9.4"
psycopg2 = "^2.9.5"
python = "^3.11"
tzdata = "^2022.7"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
whitenoise = {extras = ["brotli"], version = "^6.2.0"}

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
celery-stubs = "^0.1.2"
django-stubs = "*"
djhtml = "*"
flakeheaven = "*"
isort = "*"
mypy = "*"
mypy-extensions = "*"
myst-parser = "*"
pre-commit = "*"
pydocstyle = {version = "*", extras = ["toml"]}
pytest = "*"
pytest-django = "*"
pytest-mock = "*"
pyupgrade = "*"
sphinx = "*"
sphinx_rtd_theme = "*"
toml-sort = "*"
watchdog = "^2.2.0"
werkzeug = "*"

[tool.pydocstyle]
add_ignore = "D100,D104,D105,D106,D107,D200,D203,D205,D212,D400"
convention = "google"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "{{ cookiecutter.package_name }}.settings"
addopts = "--durations=10"
markers = [
  'api: mark a test as being an API test',
  'integration: mark a test as being an integration test with external dependencies'
]
norecursedirs = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  "__pycache__",
  "node_modules",
  "static",
  "static-files"
]
python_files = [
  "tests.py",
  "test_*.py",
  "*_tests.py"
]
